// src/app/interfaces/IMovie.ts
export interface IMovie {
    id?: string;
    title?: string;
    linkPage?: string;
    downloadLinks?: any[];
    descriptions?: string;
    imageUrl?: string;
    imdbInfo?: any;
    timestamp?: number;
    interested?: string;
    originalTitle?: string;
    tags?: string[];
    favorite?: string;
    watched?: string;
    trailer?: string;
}